Time: 0 ms (100.00%) | Memory: 7.7 MB (87.67%) - LeetSync

+++ 367-valid-perfect-square/valid-perfect-square.cpp
@@ -5,7 +5,7 @@ public:
         int low=0,high=num;
         while( low <= high)
         {
-            long long mid = low + (high-low)/2;
+            long mid = low + (high-low)/2;
             if(mid*mid == num )
             {
                 return true;
